+++ credativ_pg_migrator/connectors/postgresql_connector.py
@@ -1602,6 +1602,8 @@ class PostgreSQLConnector(DatabaseConnector):
         if self.on_error_action == 'stop':
             self.config_parser.print_log_message('ERROR', "postgresql_connector: handle_error: Stopping due to error.")
             exit(1)
+        else:
+            self.config_parser.print_log_message('WARNING', f"postgresql_connector: handle_error: Error caught, but continuing as requested by configuration (on_error_action='{self.on_error_action}').")
 
     def fetch_sequences(self, schema_name: str):
         return {}