+++ credativ_pg_migrator/planner.py
@@ -1247,6 +1247,8 @@ class Planner:
         if self.on_error_action == 'stop':
             self.config_parser.print_log_message('ERROR', "planner: handle_error: Stopping due to error.")
             exit(1)
+        else:
+            self.config_parser.print_log_message('WARNING', f"planner: handle_error: Error caught, but continuing as requested by configuration (on_error_action='{self.on_error_action}').")
 
     def check_pausing_resuming(self):
         if self.config_parser.pause_migration_fired():